+++ tool-openssl/rsa_test.cc
@@ -149,6 +149,23 @@ protected:
     ASSERT_TRUE(PEM_write_RSAPrivateKey(in_file.get(), rsa.get(), nullptr, nullptr, 0, nullptr, nullptr));
   }
 
+  void RunCommandsAndCompareOutput(const std::string &tool_command, const std::string &openssl_command) {
+    int tool_result = system(tool_command.c_str());
+    ASSERT_EQ(tool_result, 0) << "AWS-LC tool command failed: " << tool_command;
+
+    int openssl_result = system(openssl_command.c_str());
+    ASSERT_EQ(openssl_result, 0) << "OpenSSL command failed: " << openssl_command;
+
+    std::ifstream tool_output(out_path_tool);
+    this->tool_output_str = std::string((std::istreambuf_iterator<char>(tool_output)), std::istreambuf_iterator<char>());
+    std::ifstream openssl_output(out_path_openssl);
+    this->openssl_output_str = std::string((std::istreambuf_iterator<char>(openssl_output)), std::istreambuf_iterator<char>());
+
+    std::cout << "AWS-LC tool output:" << std::endl << this->tool_output_str << std::endl;
+    std::cout << "OpenSSL output:" << std::endl << this->openssl_output_str << std::endl;
+  }
+
+
   void TearDown() override {
     if (tool_executable_path != nullptr && openssl_executable_path != nullptr) {
       RemoveFile(in_path);
@@ -166,6 +183,29 @@ protected:
   std::string tool_output_str;
   std::string openssl_output_str;
 };
+
+// Helper function to trim whitespace from both ends of a string to test RSA output
+static inline std::string &trim(std::string &s) {
+  s.erase(s.begin(), std::find_if(s.begin(), s.end(), [](unsigned char ch) {
+      return !std::isspace(static_cast<unsigned char>(ch));
+  }));
+  s.erase(std::find_if(s.rbegin(), s.rend(), [](unsigned char ch) {
+      return !std::isspace(static_cast<unsigned char>(ch));
+  }).base(), s.end());
+  return s;
+}
+
+// Helper function to read file content into a string
+std::string ReadFileToString(const std::string& file_path) {
+  std::ifstream file_stream(file_path, std::ios::binary);
+  if (!file_stream) {
+    return "";
+  }
+  std::ostringstream buffer;
+  buffer << file_stream.rdbuf();
+  return buffer.str();
+}
+
 
 // RSA boundaries
 const std::string RSA_BEGIN = "-----BEGIN RSA PRIVATE KEY-----";
@@ -186,7 +226,7 @@ TEST_F(RSAComparisonTest, RSAToolCompareModulusOpenSSL) {
   std::string tool_command = std::string(tool_executable_path) + " rsa -in " + in_path + " > " + out_path_tool;
   std::string openssl_command = std::string(openssl_executable_path) + " rsa -in " + in_path + " > " + out_path_openssl;
 
-  RunCommandsAndCompareOutput(tool_command, openssl_command, out_path_tool, out_path_openssl, tool_output_str, openssl_output_str);
+  RunCommandsAndCompareOutput(tool_command, openssl_command);
 
   trim(tool_output_str);
   ASSERT_TRUE(CheckBoundaries(tool_output_str, RSA_BEGIN, RSA_END, BEGIN, END));
@@ -201,7 +241,7 @@ TEST_F(RSAComparisonTest, RSAToolCompareModulusNooutOpenSSL) {
   std::string tool_command = std::string(tool_executable_path) + " rsa -in " + in_path + " -modulus -noout > " + out_path_tool;
   std::string openssl_command = std::string(openssl_executable_path) + " rsa -in " + in_path + " -modulus -noout > " + out_path_openssl;
 
-  RunCommandsAndCompareOutput(tool_command, openssl_command, out_path_tool, out_path_openssl, tool_output_str, openssl_output_str);
+  RunCommandsAndCompareOutput(tool_command, openssl_command);
 
   ASSERT_EQ(tool_output_str, openssl_output_str);
 }
@@ -212,7 +252,7 @@ TEST_F(RSAComparisonTest, RSAToolCompareModulusOutOpenSSL) {
   std::string tool_command = std::string(tool_executable_path) + " rsa -in " + in_path + " -modulus -out " + out_path_tool;
   std::string openssl_command = std::string(openssl_executable_path) + " rsa -in " + in_path + " -modulus -out " + out_path_openssl;
 
-  RunCommandsAndCompareOutput(tool_command, openssl_command, out_path_tool, out_path_openssl, tool_output_str, openssl_output_str);
+  RunCommandsAndCompareOutput(tool_command, openssl_command);
 
   ScopedFILE tool_out_file(fopen(out_path_tool, "rb"));
   ASSERT_TRUE(tool_out_file);
@@ -234,7 +274,7 @@ TEST_F(RSAComparisonTest, RSAToolCompareModulusOutNooutOpenSSL) {
   std::string tool_command = std::string(tool_executable_path) + " rsa -in " + in_path + " -modulus -out " + out_path_tool + " -noout";
   std::string openssl_command = std::string(openssl_executable_path) + " rsa -in " + in_path + " -modulus -out " + out_path_openssl + " -noout";
 
-  RunCommandsAndCompareOutput(tool_command, openssl_command, out_path_tool, out_path_openssl, tool_output_str, openssl_output_str);
+  RunCommandsAndCompareOutput(tool_command, openssl_command);
 
   ScopedFILE tool_out_file(fopen(out_path_tool, "rb"));
   ASSERT_TRUE(tool_out_file);
